Completed draft of 'All the Flowers Are For Me'
Initially, Jekyll would not generate this post.

Found the following StackOverflow post that fixed the problem:

https://stackoverflow.com/questions/30625044/jekyll-post-not-generated

Jekyll does not like semicolons in the title, and dates can't be in the
future if the config.yml is not set to accept that.  Made changes and
Jekyll started to generate files into the _site folder.

Closes Issue #12.