Added future: true to config to solve posts not publishing problem
Will need to test by creating a new post (see Issue #21).

This solution is a little bit of a hack -- if it works, it should make
Jekyll publish posts to the _site folder even if their publish date
is in the future.

I did not see any settings in _config.yml that let you explicitly
set the time zone for your site. However, the front matter
I have been using in my posts (which I copied from the example repo)
give the time zone as -0400, which is Eastern Standard Time (maybe
DST comes into it?)

Responds to Issue #18.