Updated post about Heather and tested new config settings
Setting future: true in _config.yml appears to solve the problem
where if I had a post with the current day and time in the front
matter of the post, Jekyll would not publish it.

Post also updated with proper filename convention (needs date in
yyyy-mm-dd format in front of post name to work).

Closes Issue #18
Closes Issue #21