Testing upload of new post after Jekyll rebuild
Had to do extensive rebuilding of local Jekyll setup.

This commit contains a package-lock.json file that represents
fresh installs of Ruby and resulting installs of the `bundler` and
`jekyll` gems. Right now, `jekyll serve` is working and
a new post is visible on `localhost:4000` and posts pick up the
appropriate theme/formatting.

Responds to Issue #46.